+++ tests/functional/test_wps_provider.py
@@ -233,3 +233,80 @@ class WpsProviderTest(WpsConfigBase):
             with open(output_path, mode="r", encoding="utf-8") as out_file:
                 data = out_file.read()
             assert data == ncdump_data
+
+    @mocked_remote_server_requests_wps1([
+        resources.TEST_REMOTE_SERVER_URL,
+        resources.TEST_HUMMINGBIRD_WPS1_GETCAP_XML,
+        [resources.TEST_HUMMINGBIRD_DESCRIBE_WPS1_XML],
+    ])
+    def test_register_describe_execute_ncdump_no_default_format(self, mock_responses):
+        """
+        Test a remote :term:`WPS` provider that defines an optional input of ``ComplexData`` type with a default format.
+
+        When the ``ComplexData`` input is omitted (allowed by ``minOccurs=0``), the execution parsing must **NOT**
+        inject an input due to the detection of the default **format** (rather than default data/reference). If an
+        input is erroneously injected, :mod:`pywps` tends to auto-generate an empty file (from the storage linked to
+        the input), which yields an explicitly provided empty string value as ``ComplexData`` input.
+
+        An example of such process is the ``ncdump`` that
+        takes [0-100] ``ComplexData`` AND/OR [0-100] NetCDF OpenDAP URL strings as ``LiteralData``.
+        """
+        self.service_store.clear_services()
+
+        # register the provider
+        remote_provider_name = "test-wps-remote-provider-hummingbird"
+        path = "/providers"
+        data = {"id": remote_provider_name, "url": resources.TEST_REMOTE_SERVER_URL}
+        resp = self.app.post_json(path, params=data, headers=self.json_headers)
+        assert resp.status_code == 201
+
+        exec_path = f"{self.app_url}/providers/{remote_provider_name}/processes/ncdump/execution"
+        exec_file = "http://localhost.com/dont/care.nc"
+        exec_body = {
+            "mode": ExecuteMode.ASYNC,
+            "response": ExecuteResponse.DOCUMENT,
+            "inputs": [{"id": "dataset_opendap", "data": exec_file}],
+            "outputs": [{"id": "output", "transmissionMode": ExecuteTransmissionMode.VALUE}]
+        }
+        status_url = f"{resources.TEST_REMOTE_SERVER_URL}/status.xml"
+        output_url = f"{resources.TEST_REMOTE_SERVER_URL}/output.txt"
+        with open(resources.TEST_HUMMINGBIRD_STATUS_WPS1_XML, mode="r", encoding="utf-8") as status_file:
+            status = status_file.read().format(
+                TEST_SERVER_URL=resources.TEST_REMOTE_SERVER_URL,
+                LOCATION_XML=status_url,
+                OUTPUT_FILE=output_url,
+            )
+
+        ncdump_data = "Fake NetCDF Data"
+        with contextlib.ExitStack() as stack_exec:
+            # mock direct execution bypassing celery
+            for mock_exec in mocked_execute_celery():
+                stack_exec.enter_context(mock_exec)
+            # mock responses expected by "remote" WPS-1 Execute request and relevant documents
+            mock_responses.add("GET", exec_file, body=ncdump_data, headers={"Content-Type": ContentType.APP_NETCDF})
+            mock_responses.add("POST", resources.TEST_REMOTE_SERVER_URL, body=status, headers=self.xml_headers)
+            mock_responses.add("GET", status_url, body=status, headers=self.xml_headers)
+            mock_responses.add("GET", output_url, body=ncdump_data, headers={"Content-Type": ContentType.TEXT_PLAIN})
+
+            # add reference to specific provider class 'dispatch' to validate it was called with expected inputs
+            # (whole procedure must run even though a lot of parts are mocked)
+            # use the last possible method before sendoff to WPS, since filtering of omitted defaults can happen
+            # at the very last moment to accommodate for CWL needing 'null' inputs explicitly in jobs submission
+            real_wps1_process_dispatch = Wps1Process.dispatch
+            handle_wps1_process_dispatch = stack_exec.enter_context(
+                mock.patch.object(Wps1Process, "dispatch", side_effect=real_wps1_process_dispatch, autospec=True)
+            )  # type: MockPatch
+
+            # launch job execution and validate
+            resp = mocked_sub_requests(self.app, "post_json", exec_path, timeout=5,
+                                       data=exec_body, headers=self.json_headers, only_local=True)
+            assert resp.status_code in [200, 201], f"Failed with: [{resp.status_code}]\nReason:\n{resp.json}"
+            assert handle_wps1_process_dispatch.called, "WPS-1 handler should have been called by CWL runner context"
+
+            wps_exec_params = handle_wps1_process_dispatch.call_args_list[0].args
+            wps_exec_inputs = wps_exec_params[1]
+            assert isinstance(wps_exec_inputs, list), "WPS Inputs should have been passed for dispatch as 1st argument."
+            wps_exec_inputs = dict(wps_exec_inputs)
+            assert "dataset_opendap" in wps_exec_inputs, "Explicitly provided WPS inputs should be present."
+            assert "dataset" not in wps_exec_inputs, "Omitted WPS input should not be injected in the request."
+            assert wps_exec_inputs["dataset_opendap"] == exec_file